requirements.txt
asgiref==3.9.1
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
dotenv==0.9.9
pip==24.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3